pyproject.toml
[project]
name = "datadog-incident-manager"
version = "0.1.0"
description = "Autonomous incident management system using Claude Agent SDK and Datadog"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
